refactor: move error conversion traits into using place.
